Fix const, imports, docs, layout JC-567
