+++ apps/angular/src/shared/components/select-with-create/select-with-create.component.html
@@ -1,5 +1,4 @@
-<mat-select
-	[formControl]="formControl"
+<mat-select [formControl]="formControl"
 	#itemsSelect 
 	multiple
 	(openedChange)="handleSelectOpenedChange(!$event)">
